requirements.txt
opencv-python
zxing
pytesseract
Pillow
PyQt6
auto-py-to-exe 